+++ distributed/shuffle/tests/test_comm_buffer.py
@@ -1,17 +1,13 @@
 from __future__ import annotations
 
 import asyncio
-import concurrent.futures
 import math
 from collections import defaultdict
 
 import pytest
-from tornado.ioloop import IOLoop
 
-from distributed.shuffle._multi_comm import MultiComm
+from distributed.shuffle._comms import CommShardsBuffer
 from distributed.utils_test import gen_test
-
-# pytest.skip(allow_module_level=True)
 
 
 @gen_test()
@@ -21,9 +17,9 @@ async def test_basic(tmp_path):
     async def send(address, shards):
         d[address].extend(shards)
 
-    mc = MultiComm(send=send, loop=IOLoop.current())
-    mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
-    mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
+    mc = CommShardsBuffer(send=send)
+    await mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
+    await mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
 
     await mc.flush()
 
@@ -38,17 +34,16 @@ async def test_exceptions(tmp_path):
     async def send(address, shards):
         raise Exception(123)
 
-    mc = MultiComm(send=send, loop=IOLoop.current())
-    mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
+    mc = CommShardsBuffer(send=send)
+    await mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
 
     while not mc._exception:
         await asyncio.sleep(0.1)
 
     with pytest.raises(Exception, match="123"):
-        mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
+        await mc.put({"x": [b"0" * 1000], "y": [b"1" * 500]})
 
-    with pytest.raises(Exception, match="123"):
-        await mc.flush()
+    await mc.flush()
 
     await mc.close()
 
@@ -65,29 +60,27 @@ async def test_slow_send(tmpdir):
         d[address].extend(shards)
         sending_first.set()
 
-    mc = MultiComm(send=send, loop=IOLoop.current())
-    mc.max_connections = 1
-    mc.put({"x": [b"0"], "y": [b"1"]})
-    mc.put({"x": [b"0"], "y": [b"1"]})
+    mc = CommShardsBuffer(send=send, concurrency_limit=1)
+    await mc.put({"x": [b"0"], "y": [b"1"]})
+    await mc.put({"x": [b"0"], "y": [b"1"]})
     flush_task = asyncio.create_task(mc.flush())
     await sending_first.wait()
     block_send.clear()
 
     with pytest.raises(RuntimeError):
-        mc.put({"x": [b"2"], "y": [b"2"]})
+        await mc.put({"x": [b"2"], "y": [b"2"]})
         await flush_task
 
     assert [b"2" not in shard for shard in d["x"]]
 
 
 def gen_bytes(percentage: float) -> bytes:
-    num_bytes = int(math.floor(percentage * MultiComm.memory_limit))
+    num_bytes = int(math.floor(percentage * CommShardsBuffer.memory_limit))
     return b"0" * num_bytes
 
 
-@pytest.mark.parametrize("explicit_flush", [True, False])
 @gen_test()
-async def test_concurrent_puts(explicit_flush):
+async def test_concurrent_puts():
     d = defaultdict(list)
 
     async def send(address, shards):
@@ -97,24 +90,20 @@ async def test_concurrent_puts(explicit_flush):
     nshards = 10
     nputs = 20
     payload = {x: [gen_bytes(frac)] for x in range(nshards)}
-    with concurrent.futures.ThreadPoolExecutor(
-        2, thread_name_prefix="test IOLoop"
-    ) as tpe:
-        async with MultiComm(send=send, loop=IOLoop.current()) as mc:
-            loop = asyncio.get_running_loop()
-            futs = [loop.run_in_executor(tpe, mc.put, payload) for _ in range(nputs)]
 
-            await asyncio.gather(*futs)
-            if explicit_flush:
-                await mc.flush()
+    async with CommShardsBuffer(send=send) as mc:
+        futs = [asyncio.create_task(mc.put(payload)) for _ in range(nputs)]
 
-                assert not mc.shards
-                assert not mc.sizes
+        await asyncio.gather(*futs)
+        await mc.flush()
 
         assert not mc.shards
         assert not mc.sizes
-        assert len(d) == 10
-        assert sum(map(len, d[0])) == len(gen_bytes(frac)) * nputs
+
+    assert not mc.shards
+    assert not mc.sizes
+    assert len(d) == 10
+    assert sum(map(len, d[0])) == len(gen_bytes(frac)) * nputs
 
 
 @gen_test()
@@ -134,15 +123,14 @@ async def test_concurrent_puts_error():
     nshards = 10
     nputs = 20
     payload = {x: [gen_bytes(frac)] for x in range(nshards)}
-    with concurrent.futures.ThreadPoolExecutor(
-        2, thread_name_prefix="test IOLoop"
-    ) as tpe:
-        async with MultiComm(send=send, loop=IOLoop.current()) as mc:
-            loop = asyncio.get_running_loop()
-            futs = [loop.run_in_executor(tpe, mc.put, payload) for _ in range(nputs)]
 
-            with pytest.raises(OSError, match="error during send"):
-                await asyncio.gather(*futs)
+    async with CommShardsBuffer(send=send) as mc:
+        futs = [asyncio.create_task(mc.put(payload)) for _ in range(nputs)]
+
+        await asyncio.gather(*futs)
+        await mc.flush()
+        with pytest.raises(OSError, match="error during send"):
+            mc.raise_on_exception()
 
     assert not mc.shards
     assert not mc.sizes
